updates and renaming
